requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alembic==0.9.6
bcrypt==3.1.4
certifi==2017.11.5        # via requests
cffi==1.11.2              # via bcrypt
chardet==3.0.4            # via requests
click==6.7                # via flask
configparser==3.5.0
docutils==0.14            # via python-daemon
flask-login==0.4.0
flask==0.12.2
future==0.16.0
idna==2.6                 # via requests
ipy==0.83
itsdangerous==0.24        # via flask
jinja2==2.10              # via flask
lockfile==0.12.2          # via python-daemon
mako==1.0.7               # via alembic
markupsafe==1.0           # via jinja2, mako
netifaces==0.10.6
olefile==0.44             # via pillow
path.py==10.5
pillow==4.3.0
psutil==5.4.1
pyasn1==0.3.7             # via pysmb
pycparser==2.18           # via cffi
pycrypto==2.6.1
pymediainfo==2.2.0
pymysql==0.7.11
pysmb==1.1.22
python-daemon==2.1.2
python-dateutil==2.6.1    # via alembic
python-editor==1.0.3      # via alembic
pytvmaze==2.0.8
requests==2.18.4
roman==2.0.0
six==1.11.0
sqlalchemy==1.1.15
tmdbsimple==1.8.0
tvdbsimple==1.0.4
unidecode==0.4.21
urllib3==1.22             # via requests
werkzeug==0.12.2          # via flask
xmltodict==0.11.0
